* C++/YellowBelt/1/blocks (Масса блоков): исправлено

+++ C++/YellowBelt/1/blocks/blocks.cpp
@@ -4,13 +4,13 @@ using namespace std;
 
 int main() {
   int n, r;
-  int64_t sum = 0;
+  uint64_t sum = 0;
   cin >> n >> r;
   for (int i = 0; i < n; ++i) {
     int w, h, d;
     cin >> w >> h >> d;
-    int64_t v = static_cast<int64_t>(w) * h * d;
-    int64_t m = v * r;
+    uint64_t v = static_cast<uint64_t>(w) * h * d;
+    uint64_t m = static_cast<uint64_t>(v) * r;
     sum += m;
   }
